Change in QuadraticEqSolver and its launcher
Changed some primitives from int to double so
to allow equations like 1.4x^2 +5x + 1 = 0

+++ random/QuadraticEqSolver.java
@@ -2,30 +2,30 @@ package random;
 //This class will try to solve a quadratic equation
 //As the input, user will provide the coefficients of the terms and the constant term
 class QuadraticEqSolver {
-    private int a, b, c;
+    private double a, b, c;
     private double xValue1, xValue2, discriminant;
     
     //Setters
-    public void setA(int a) {
+    public void setA(double a) {
         if (a != 0)
             this.a = a;
     }
-    public void setB(int b) {
+    public void setB(double b) {
         this.b = b;
     }
-    public void setC(int c) {
+    public void setC(double c) {
         this.c = c;
     }
     
     //I don't think that there is a need of getters
     
-    public double discriminantFinder (int a, int b, int c){
+    public double discriminantFinder (double a, double b, double c){
         discriminant = b*b - 4*a*c;
         return discriminant;
     }
     
     //Last method to totally solve the equation
-    public String solveForX (int b, double discriminant, int a, boolean hasOneRoot){
+    public String solveForX (double b, double discriminant, double a, boolean hasOneRoot){
         //We will be needing one value at least
         xValue1 = (Math.sqrt (discriminant) - b) / (2*a);
